refactor: 카드 포맷팅 함수 util로 분리
