+++ hooks/src/lib/useCardNumber.ts
@@ -1,6 +1,7 @@
 import { useState } from 'react';
 import { REGEX, META_CARD, CardBrand } from './constants';
 import useCardType from './useCardType';
+import { formatCardNumber } from './../utils/formatCardNumber';
 
 interface CardNumberInfo {
   cardNumber: string[];
@@ -18,20 +19,6 @@ const useCardNumber = () => {
   });
   const { handleCardTypeChange } = useCardType();
 
-  const formatCardNumber = (cardNumber: string, format: number[]) => {
-    const result: string[] = [];
-    let index = 0;
-    for (const length of format) {
-      if (index + length >= cardNumber.length) {
-        result.push(cardNumber.slice(index));
-        break;
-      }
-      result.push(cardNumber.slice(index, index + length));
-      index += length;
-    }
-    return result;
-  };
-
   const handleCardNumberChange = (cardNumber: string) => {
     const cardNumberWithoutSpace = cardNumber.replace(/\D/g, '');
     const newCardType = handleCardTypeChange(cardNumberWithoutSpace);
@@ -39,9 +26,11 @@ const useCardNumber = () => {
     const errors: string[] = [];
     const isNumeric = REGEX.onlyNumber.test(cardNumberWithoutSpace);
     const isValidNumber = cardNumberWithoutSpace.length === META_CARD[newCardType].maxLength;
+
     if (!isNumeric) {
       errors.push('숫자를 입력해주세요.');
     }
+
     if (!isValidNumber) {
       errors.push(`${META_CARD[newCardType].maxLength}자리 숫자를 입력해주세요.`);
     }
